Move dependency struct to dependencies module
