+++ src/project/mod.rs
@@ -4,7 +4,7 @@ pub mod manifests;
 pub mod packager;
 pub mod settings;
 
-use crate::project::dependencies::DependencyType;
+use crate::project::dependencies::{Dependency, DependencyType};
 use crate::project::manifests::{DependenciesManifest, Manifest, VersionManifest};
 use crate::project::settings::ProjectSettings;
 use crate::util;
@@ -23,21 +23,6 @@ pub struct Server {
     pub source: String,
     pub brand: String,
     pub version: String,
-}
-
-// Workaround for https://github.com/serde-rs/serde/issues/368
-pub const fn default_bool<const V: bool>() -> bool {
-    V
-}
-
-#[derive(Serialize, Deserialize, Default, Debug, Clone)]
-pub struct Dependency {
-    pub source: Option<String>,
-    pub version: Option<String>,
-    #[serde(rename = "type", default)]
-    pub dependency_type: DependencyType,
-    #[serde(default = "default_bool::<false>")]
-    pub required: bool,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
